code to run more iterations of the simulation

+++ scenarios/v06.py
@@ -277,12 +277,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path_burn_in_movement):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path_burn_in_movement, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path_burn_in_movement):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path_burn_in_movement, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 
 ###################################################
@@ -394,12 +394,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path_undetected_spread):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path_undetected_spread, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path_undetected_spread):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path_undetected_spread, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 
 ###################################################
@@ -454,12 +454,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path_first_report):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path_first_report, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path_first_report):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path_first_report, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 
 ###################################################
@@ -528,13 +528,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
-
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 ###################################################
 # ---- Run second set of actions ------------------#
@@ -602,13 +601,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
-
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 ###################################################
 # ---- Run third set of actions ------------------#
@@ -676,14 +674,12 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
-
-exit(0)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 ###################################################
 # ---- Run fourth set of actions ------------------#
@@ -751,16 +747,16 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
 
 
 ###################################################
-# ---- Run fourth set of actions ------------------#
+# ---- Run fifth set of actions ------------------#
 ###################################################
 
 # # generate a list of scheduled management actions
@@ -825,9 +821,89 @@ download_folder_path = os.path.join(folder_path_main, "download_" + unique_outpu
 if not os.path.exists(download_folder_path):
     os.makedirs(download_folder_path)
 
-# Loop through the files in the source directory and copy just the png or csv files
-for file in os.listdir(folder_path):
-    if file.endswith(".png") or file.endswith(".csv"):
-        source_path = os.path.join(folder_path, file)
-        destination_path = os.path.join(download_folder_path, file)
-        shutil.copy(source_path, destination_path)
+    # Loop through the files in the source directory and copy just the png or csv files
+    for file in os.listdir(folder_path):
+        if file.endswith(".png") or file.endswith(".csv"):
+            source_path = os.path.join(folder_path, file)
+            destination_path = os.path.join(download_folder_path, file)
+            shutil.copy(source_path, destination_path)
+
+
+###################################################
+# ---- Running further actions ------------------#
+###################################################
+
+# # generate a list of scheduled management actions
+# # actions, basic: date, property_id, action-to-take-on-date, extra deets for action if necessary (e.g., if culling, the number of animals culled on that day)
+
+max_runs = 10
+next_run = 6
+while next_run <= max_runs:
+
+    actions_input = os.path.join(folder_path_main, f"actions_{next_run}.xlsx")
+    property_jobs = pd.read_excel(actions_input, sheet_name="jobs")
+    zones_based_jobs = pd.read_excel(actions_input, sheet_name="zone_jobs")
+    property_based_zones = pd.read_excel(actions_input, sheet_name="zones")  # could consider "expanding to SAL, LGA" or something like that
+    days_to_run_for = 3  # running for three days feels better
+
+    number = next_run + 3
+    unique_output = f"{number:02d}_actions_{next_run}"
+    folder_path = os.path.join(folder_path_main, unique_output)
+
+    if not os.path.exists(folder_path):
+        os.makedirs(folder_path)
+
+    spread_properties_filename = os.path.join(folder_path, "properties_" + unique_output)
+    spread_diseaseoutbreak_filename = os.path.join(folder_path, "outbreakobject_" + unique_output)
+
+    random.seed(215)
+    np.random.seed(216)
+    if not os.path.exists(spread_properties_filename) or not os.path.exists(spread_diseaseoutbreak_filename):
+        # adjust the plotting parameters for this new scenario
+        diseaseoutbreak.set_plotting_parameters(
+            xlims=xlims,
+            ylims=ylims,
+            plotting=True,
+            folder_path=folder_path,
+            unique_output=unique_output,
+        )
+
+        properties, movement_records, time, total_culled_animals, job_manager = diseaseoutbreak.simulate_HPAI_outbreak_management(
+            properties, property_jobs, zones_based_jobs, property_based_zones, days_to_run_for
+        )
+
+        # and then resave the end state
+        with open(spread_properties_filename, "wb") as file:
+            pickle.dump(properties, file)
+
+        # and save the diseaseoutbreak object
+        with open(spread_diseaseoutbreak_filename, "wb") as file:
+            pickle.dump(diseaseoutbreak, file)
+
+        total_infected = 0
+        for property_i in properties:
+            if property_i.exposure_date != "NA":
+                total_infected += 1
+
+        print(f"Total number of infected premises: {total_infected}")
+    else:
+        with open(spread_properties_filename, "rb") as file:
+            properties = pickle.load(file)
+        with open(spread_diseaseoutbreak_filename, "rb") as file:
+            diseaseoutbreak = pickle.load(file)
+
+    HPAI_functions.save_approx_known_data(properties, folder_path, unique_output)
+
+    download_folder_path = os.path.join(folder_path_main, "download_" + unique_output)
+
+    if not os.path.exists(download_folder_path):
+        os.makedirs(download_folder_path)
+
+        # Loop through the files in the source directory and copy just the png or csv files
+        for file in os.listdir(folder_path):
+            if file.endswith(".png") or file.endswith(".csv"):
+                source_path = os.path.join(folder_path, file)
+                destination_path = os.path.join(download_folder_path, file)
+                shutil.copy(source_path, destination_path)
+
+    next_run += 1
